Extract the logging stuff into its own file.
