+++ local-modules/app-setup/Application.js
@@ -2,11 +2,9 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import chalk from 'chalk';
 import express from 'express';
 import express_ws from 'express-ws';
 import fs from 'fs';
-import morgan from 'morgan';
 import path from 'path';
 
 import { PostConnection, TargetMap, WsConnection } from 'api-server';
@@ -18,6 +16,7 @@ import { Dirs } from 'server-env';
 
 import Authorizer from './Authorizer';
 import DebugTools from './DebugTools';
+import RequestLogger from './RequestLogger';
 
 /** Logger. */
 const log = new SeeAll('app');
@@ -70,101 +69,12 @@ export default class Application {
    * Sets up logging for webserver requests.
    */
   _addRequestLogging() {
-    const app = this._app;
-
     // Stream to write to, when logging to a file.
-    const logStream = fs.createWriteStream(
+    const accessStream = fs.createWriteStream(
       path.resolve(Dirs.VAR_DIR, 'access.log'),
       {flags: 'a'});
 
-    // These log regular (non-websocket) requests at the time of completion,
-    // including a short colorized form to the console and a longer form to a
-    // file.
-
-    // Returns whether (`true`) or not (`false`) the given request is a
-    // websocket request.
-    function isWebsocketRequest(req) {
-      // Case doesn't matter, hence the regex test instead of just `===`.
-      const upgrade = req.get('upgrade');
-      return (upgrade !== undefined) && upgrade.match(/^websocket$/i);
-    }
-
-    // Opposite of `isWebsocketRequest()`, used for log skipping.
-    function notWebsocketRequest(req) {
-      return !isWebsocketRequest(req);
-    }
-
-    // This is a status-aware logger, roughly based on morgan's built-in `dev`
-    // style. If `res` is passed as `null`, this indicates that the logging is
-    // being done at the start of the request, so (e.g.) things like content
-    // length won't be available.
-    function shortColorLog(req, res) {
-      const status      = (res === null) ? 0 : (res.statusCode || 0);
-      const statusStr   = `${status || '-'}  `.slice(0, 3);
-      const colorFn     = Application._colorForStatus(status);
-      const isWebsocket = isWebsocketRequest(req);
-      const method      = `${isWebsocket ? 'WS' : req.method}   `.slice(0,4);
-
-      console.log('===== a');
-      // `express-ws` appends a pseudo-path `/.websocket` to the end of
-      // websocket requests. We chop that off here.
-      const url = isWebsocket
-        ? req.originalUrl.replace(/\/\.websocket$/, '')
-        : req.originalUrl;
-
-      let contentLength = (res === null) ? undefined : res.get('content-length');
-      if (contentLength === undefined) {
-        contentLength = '-';
-      } else if (contentLength > (1024 * 1024)) {
-        contentLength = `${Math.round(contentLength / 1024 / 1024 * 10) / 10}M`;
-      } else if (contentLength > 1024) {
-        contentLength = `${Math.round(contentLength / 1024 * 10) / 10}K`;
-      } else {
-        contentLength = `${contentLength}B`;
-      }
-
-      if (contentLength.length < 7) {
-        contentLength += ' '.repeat(7 - contentLength.length);
-      }
-
-      console.log('===== z');
-      return `${colorFn(statusStr)} ${contentLength} ${method} ${url}`;
-    }
-
-    app.use(morgan(
-      (tokens_unused, req, res) => { return shortColorLog(req, res); },
-      {
-        stream: log.infoStream
-      }
-    ));
-
-    app.use(morgan(
-      'common',
-      {
-        stream: logStream
-      }
-    ));
-
-    // These log websocket requests, at the time of request start (not at the
-    // time of completion because these are long-lived requests).
-
-    app.use(morgan(
-      (tokens_unused, req, res_unused) => { return shortColorLog(req, null); },
-      {
-        stream:    log.infoStream,
-        immediate: true,
-        skip:      notWebsocketRequest
-      }
-    ));
-
-    app.use(morgan(
-      'common',
-      {
-        stream:    logStream,
-        immediate: true,
-        skip:      notWebsocketRequest
-      }
-    ));
+    RequestLogger.addLoggers(this._app, log.infoStream, accessStream);
   }
 
   /**
@@ -202,20 +112,5 @@ export default class Application {
   _addDevModeRoutes() {
     const app = this._app;
     app.use('/debug', new DebugTools(this._doc).requestHandler);
-  }
-
-  /**
-   * Given an HTTP status, returns the corresponding `chalk` coloring function,
-   * or a no-op function if the status doesn't demand colorization.
-   *
-   * @param {number} status The HTTP status code.
-   * @returns {function} The corresponding coloring function.
-   */
-  static _colorForStatus(status) {
-    if      (status >= 500) { return chalk.red;    }
-    else if (status >= 400) { return chalk.yellow; }
-    else if (status >= 300) { return chalk.cyan;   }
-    else if (status >= 200) { return chalk.green;  }
-    else                    { return ((x) => x);   } // No-op by default.
   }
 }
